+++ src/components/config.ts
@@ -1,0 +1,3 @@
+export const BASE_URL = 'https://wofp.jasperagrante.com'
+// export const BASE_URL = 'http://localhost:8787'
+export const SYNC_INTERVAL = 60 * 60 * 1000